fix: generate types from server plugin (#3366)
