+++ build.config.ts
@@ -2,13 +2,5 @@ import { defineBuildConfig } from 'unbuild'
 
 export default defineBuildConfig({
   entries: ['src/module'],
-  externals: [
-    'node:fs',
-    'node:url',
-    'webpack',
-    '@babel/parser',
-    'unplugin-vue-router',
-    'unplugin-vue-router/options',
-    'chokidar'
-  ]
+  externals: ['node:fs', 'node:url', 'webpack', '@babel/parser', 'unplugin-vue-router', 'unplugin-vue-router/options']
 })
