+++ package.json
@@ -123,7 +123,6 @@
     "@unhead/vue": "^1.8.8",
     "bumpp": "^9.4.1",
     "changelogithub": "^0.13.7",
-    "chokidar": "^4.0.1",
     "consola": "^3",
     "eslint": "^9.5.0",
     "eslint-config-prettier": "^9.1.0",
