+++ src/template.ts
@@ -162,5 +162,5 @@ export const DEFAULT_DYNAMIC_PARAMS_KEY = ${JSON.stringify(DEFAULT_DYNAMIC_PARAM
 export const DEFAULT_COOKIE_KEY = ${JSON.stringify(DEFAULT_COOKIE_KEY)}
 export const SWITCH_LOCALE_PATH_LINK_IDENTIFIER = ${JSON.stringify(SWITCH_LOCALE_PATH_LINK_IDENTIFIER)}
 
-${(options.dev && options.hmr && codeHMR) || ''}`
+${(options.dev && options.hmr && !options.isServer && codeHMR) || ''}`
 }
